% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{bootpval}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{Vector containing the sample}

\item{conf.level}{The degree of confidence, equal to 1-alpha}

\item{iter}{How many iterations (resamplings) to do}

\item{mu0}{The null value of the null hypothesis}

\item{test}{The type of test to run, left tail, right tail, or both}
}
\value{
Outputs a graph, as well as the quantiles of the RAR region and the p-value in a list
}
\description{
bootpval
}
