% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Quantity representing how many samples to take}

\item{n}{Quantity representing how many items are to be in each sample}

\item{p}{Quantity representing probability that an item is positive or a success}
}
\value{
Table with ratios of results and a plot of said table
}
\description{
mybin
}
