% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(data)
}
\arguments{
\item{data}{Vector containing a sample to calculate the mean of}
}
\value{
returns a 95% confidence interval estimating the mean
}
\description{
myci
}
