% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n = 5, iter = 10000, lambda = 10, ...)
}
\arguments{
\item{n}{number of items in each sample}

\item{iter}{number of samples to take}

\item{lambda}{lambda value for poisson distribution sampling}

\item{...}{}
}
\value{
Outputs a multi-part plot showing various facets of the sampling
}
\description{
mycltp
}
