% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{Vector of sample values}

\item{mu}{Either the population mean or a sequence of values guessing at it}

\item{sig}{Either the population std deviation or a sequence of values guessing at it}

\item{...}{}
}
\value{
Calculations giving results as well as a graph
}
\description{
mymlnorm
}
