% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean value for a normal distribution}

\item{sigma}{Standard deviation for a normal distribution}

\item{a}{Value to calculate the probability of, P(Y>=a)}
}
\value{
Graph of the curve with shading based on a, as well as a calculation of P(Y>=a)
}
\description{
myncurve
}
