% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise1.R
\name{piecewise1}
\alias{piecewise1}
\title{Function for Calculating Piecewise Lines}
\usage{
piecewise1(x, coef, fulcrum)
}
\arguments{
\item{x}{A vector of quantitative data}

\item{coef}{A vector of size 3 with coefficients for the function}

\item{fulcrum}{Number indicating X value the function changes at}
}
\value{
A vector of outputs based on the piecewise function
}
\description{
Function for Calculating Piecewise Lines
}
